requirements-dev.txt
-r requirements.txt

# Flake8 deps
flake8==3.5.0
mccabe==0.6.1
pycodestyle==2.3.1
pyflakes==1.6.0

# Pytest deps
atomicwrites==1.1.5
attrs==18.1.0
more-itertools==4.2.0
pluggy==0.6.0
py==1.5.4
pytest==3.6.2
six==1.11.0
pytest-django==3.3.3

# Coverage deps
coverage==4.5.1
pytest-cov==2.5.1